Support several embassy-like workers
The AMQP setup stuff is now possibly correct.  I was fairly confused
about the concepts of queues and binding keys when I first wrote it, and
it only accidentally worked.  Now there are two exchanges, one fanout
and one direct, and we use no binding keys.

Commands are published to the direct exchange, and events are published
to the fanout exchange.  An "embassy-like worker," then, is something
that consumes events and produces commands.  Due to the fanout, all
workers receive all events.